model-creation/requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bentoml==1.4.12
cattrs==23.1.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.7
cloudpickle==3.1.1
cryptography==44.0.2
Deprecated==1.2.18
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
kantoku==0.18.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.570.86
opencv-python==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-aiohttp-client==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pip-requirements-parser==32.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
questionary==2.1.0
rich==14.0.0
schema==0.7.7
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
